move: 컴포넌트 파일 구조 정리
