+++ src/components/common/InputsUserInfo/InputProfileImage/InputProfileImage.style.ts
@@ -1,5 +1,5 @@
 import { css } from '@emotion/react';
-import { Theme } from '../../../styles/Theme';
+import { Theme } from '../../../../styles/Theme';
 
 export const profileImageSelectContainer = css({
   display: 'flex',
