+++ src/components/common/Profile/ProfileImage/ProfileImage.tsx
@@ -1,8 +1,8 @@
 import { css } from '@emotion/react';
 import type { ComponentPropsWithRef, ForwardedRef } from 'react';
 import React, { forwardRef } from 'react';
-import { Theme } from '../../../styles/Theme';
-import type { Size } from '../../../types/temp';
+import { Theme } from '../../../../styles/Theme';
+import type { Size } from '../../../../types/temp';
 import { getSizeStyling, profileImageContainer } from './ProfileImage.style';
 
 export interface ProfileImageProps extends ComponentPropsWithRef<'img'> {
