+++ src/components/mypage/ListItem/ItemVotedPosts.tsx
@@ -1,4 +1,5 @@
 import React, { ComponentPropsWithRef } from 'react';
+import { ItemVotedPostsType } from '../../../types/mypage';
 import {
   hoverStyling,
   mypageListItemContainer,
@@ -6,23 +7,22 @@ import {
   noContainer,
   withoutNoContainer,
 } from './ListItem.style';
-import { ItemCommentsType } from '../../../../types/mypage';
 
-export interface ItemCommentsProps extends ComponentPropsWithRef<'li'> {
-  item: ItemCommentsType;
+export interface ItemVotedPostsProps extends ComponentPropsWithRef<'li'> {
+  item: ItemVotedPostsType;
 }
 
-const ItemComments = ({ item }: ItemCommentsProps) => {
+const ItemVotedPosts = ({ item }: ItemVotedPostsProps) => {
   return (
     <li css={mypageListItemContainer}>
       <span css={[mypageTextStyling('xSmall'), noContainer]}>{item.id}</span>
       <div css={[withoutNoContainer, hoverStyling]}>
-        <p css={mypageTextStyling('small')}>{item.comment}</p>
+        <p css={mypageTextStyling('small')}>{item.position}</p>
         <p css={mypageTextStyling('xSmall')}>{item.date}</p>
-        <p css={[mypageTextStyling('xSmall')]}>{item.title}</p>
+        <p css={mypageTextStyling('xSmall')}>{item.title}</p>
       </div>
     </li>
   );
 };
 
-export default ItemComments;
+export default ItemVotedPosts;
