+++ src/pages/MyPage/HistoryPage/TabPage/PostsPage.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
-import PageNavigation from '../../../../components/Paginavigation/Paginavigation';
-import List from '../../sections/List/List';
-import ItemPosts from '../../sections/ListItem/ItemPosts';
+import PageNavigation from '../../../../components/common/Paginavigation/Paginavigation';
+import List from '../../../../components/mypage/List/List';
+import ItemPosts from '../../../../components/mypage/ListItem/ItemPosts';
 
 const PostsPage = () => {
   const items = [
